Separates profile gathering and challenges generation
